requirements.txt
beautifulsoup4==4.11.1
deep_translator==1.9.2
pandas==1.5.1
requests==2.28.1
toloka_kit==0.1.23
tqdm==4.64.1
